refactor(dalux): extract parseDaluxNode into its own module
From a local CodeRabbit run against the final head. Unlike the other two
~400-line findings I dismissed today, this one is mine to fix: provider.ts was
398 lines on main and the node support pushed it to 442, so my change is what
crossed the limit rather than compounding a pre-existing overrun.

parseDaluxNode and its allowlist now live in node-url.ts, bringing provider.ts
back to 406. No behaviour change; the test moves its import and nothing else.

source-dalux 152 passed / 12 skipped, relay 24/24, typecheck, lint, api-surface
and source-text gates clean.
