+++ packages/source-dalux/src/provider.ts
@@ -36,6 +36,7 @@ import type {
 } from '@ifc-lite/plugin-api';
 
 import { DALUX_MANIFEST } from './manifest.js';
+import { parseDaluxNode } from './node-url.js';
 import { BrowserDaluxApiClient, DaluxHttpError, fetchPage, fetchAllPages } from './http-client.js';
 import {
   LATEST_REVISION,
@@ -51,43 +52,6 @@ import {
 
 const DEFAULT_BASE_URL = 'https://node1.field.dalux.com/service/api';
 
-/** `node1`, `node2`, ... Mirrors the relay's own allowlist. */
-const DALUX_NODE_PATTERN = /^node[1-9][0-9]{0,2}$/;
-
-/**
- * Read a user-entered Dalux base URL and return just the node name.
- *
- * Dalux assigns each customer a node and prints the base URL beside the API
- * key, so users paste the whole thing. Only the node name is kept, and only if
- * it is a real Dalux field node: everything else about the URL is ours to
- * decide: `/api/dalux` is unauthenticated and publicly reachable, so any host
- * the relay can be aimed at becomes reachable by anyone through our egress
- * IPs. Keeping the origin ours to build bounds that to Dalux.
- *
- * Returns undefined for blank input or the default node, so the common case
- * sends no parameter at all. Throws on input that looks like a deliberate
- * attempt to reach somewhere else, because silently falling back to node1
- * would present as "my key does not work" rather than "that URL is wrong".
- */
-export function parseDaluxNode(raw: string | undefined | null): string | undefined {
-  const trimmed = (raw ?? '').trim();
-  if (trimmed === '') return undefined;
-
-  let host: string;
-  try {
-    host = new URL(trimmed.includes('://') ? trimmed : `https://${trimmed}`).hostname;
-  } catch {
-    throw new Error(`Not a valid Dalux base URL: ${trimmed}`);
-  }
-
-  const match = /^(node[1-9][0-9]{0,2})\.field\.dalux\.com$/.exec(host);
-  if (!match || !DALUX_NODE_PATTERN.test(match[1])) {
-    throw new Error(
-      `Not a Dalux node URL: ${trimmed}. Expected something like https://node2.field.dalux.com/service/api`,
-    );
-  }
-  return match[1] === 'node1' ? undefined : match[1];
-}
 
 export class DaluxBuildProvider implements FileSourceProvider {
   readonly manifest = DALUX_MANIFEST;
